+++ httpclient-vertx/src/main/java/io/fabric8/kubernetes/client/vertx/VertxHttpClientFactory.java
@@ -17,60 +17,31 @@ package io.fabric8.kubernetes.client.vertx;
 
 import io.fabric8.kubernetes.client.Config;
 import io.vertx.core.Vertx;
-import io.vertx.core.VertxOptions;
-import io.vertx.core.file.FileSystemOptions;
 import io.vertx.ext.web.client.WebClientOptions;
-
-import static io.vertx.core.spi.resolver.ResolverProvider.DISABLE_DNS_RESOLVER_PROP_NAME;
 
 public class VertxHttpClientFactory implements io.fabric8.kubernetes.client.http.HttpClient.Factory {
 
-  private static final class VertxHolder {
-
-    private static final Vertx INSTANCE = createVertxInstance();
-
-    private static synchronized Vertx createVertxInstance() {
-      // We must disable the async DNS resolver as it can cause issues when resolving the Vault instance.
-      // This is done using the DISABLE_DNS_RESOLVER_PROP_NAME system property.
-      // The DNS resolver used by vert.x is configured during the (synchronous) initialization.
-      // So, we just need to disable the async resolver around the Vert.x instance creation.
-      final String originalValue = System.getProperty(DISABLE_DNS_RESOLVER_PROP_NAME);
-      Vertx vertx;
-      try {
-        System.setProperty(DISABLE_DNS_RESOLVER_PROP_NAME, "true");
-        vertx = Vertx.vertx(new VertxOptions()
-            .setFileSystemOptions(new FileSystemOptions().setFileCachingEnabled(false).setClassPathResolvingEnabled(false))
-            .setUseDaemonThread(true));
-      } finally {
-        // Restore the original value
-        if (originalValue == null) {
-          System.clearProperty(DISABLE_DNS_RESOLVER_PROP_NAME);
-        } else {
-          System.setProperty(DISABLE_DNS_RESOLVER_PROP_NAME, originalValue);
-        }
-      }
-      return vertx;
-    }
-  }
-
-  private final Vertx vertx;
+  final Vertx sharedVertx;
 
   public VertxHttpClientFactory() {
-    this(VertxHolder.INSTANCE);
-    Runtime.getRuntime().addShutdownHook(new Thread(() -> {
-      if (vertx != null) {
-        vertx.close();
-      }
-    }));
+    this(null);
   }
 
-  public VertxHttpClientFactory(Vertx vertx) {
-    this.vertx = vertx;
+  /**
+   * Create a new instance of the factory that will reuse the provided {@link Vertx} instance.
+   * <p>
+   * It's the user's responsibility to manage the lifecycle of the provided Vert.x instance.
+   * Operations such as close, and so on are left on hands of the user.
+   *
+   * @param sharedVertx the Vertx instance to use.
+   */
+  public VertxHttpClientFactory(Vertx sharedVertx) {
+    this.sharedVertx = sharedVertx;
   }
 
   @Override
   public VertxHttpClientBuilder<VertxHttpClientFactory> newBuilder() {
-    return new VertxHttpClientBuilder<>(this, vertx);
+    return new VertxHttpClientBuilder<>(this, sharedVertx);
   }
 
   /**